+++ src/main/java/org/trainee/userservice/exception/RecordNotFoundException.java
@@ -1,7 +1,7 @@
 package org.trainee.userservice.exception;
 
 public class RecordNotFoundException extends RuntimeException {
-    public RecordNotFoundException(Integer id) {
+    public RecordNotFoundException(Long id) {
         super("Record with id=" + id + " wasn't found");
     }
 }
