+++ src/main/java/org/trainee/userservice/exception/RecordStillActiveException.java
@@ -1,7 +1,7 @@
 package org.trainee.userservice.exception;
 
 public class RecordStillActiveException extends RuntimeException {
-    public RecordStillActiveException(Integer id) {
+    public RecordStillActiveException(Long id) {
         super("Record with id=" + id + " is still active. Deactivate it first");
     }
 }
